+++ simvue/sender/base.py
@@ -45,8 +45,24 @@ class Sender:
         throw_exceptions: bool = False,
         retry_failed_uploads: bool = False,
         run_notification: typing.Literal["none", "all", "email"] = "none",
+        run_retention_period: str | None = None,
     ) -> None:
-        """Creates required local directories."""
+        """Initialise a local data sender.
+
+        Parameters
+        ----------
+        cache_directory : pydantic.DirectoryPath | None, optional
+            The directory where cached files are stored, else use default.
+        max_workers : int, optional
+            The maximum number of threads to use, default 5.
+        threading_threshold : int, optional
+            The number of cached files above which threading will be used, default 10.
+        throw_exceptions : bool, optional
+            Whether to throw exceptions as they are encountered in the sender,
+            default is False (exceptions will be logged)
+        retry_failed_uploads : bool, optional
+            Whether to retry sending objects which previously failed, by default False
+        """
         _local_config: SimvueConfiguration = SimvueConfiguration.fetch()
         self._cache_directory = cache_directory or _local_config.offline.cache
         self._cache_directory.joinpath("server_ids").mkdir(parents=True, exist_ok=True)
@@ -57,6 +73,8 @@ class Sender:
         self._lock_path = self._cache_directory.joinpath("sender.lock")
         self._thread_lock = threading.Lock()
         self._run_notification: typing.Literal["none", "email"] = run_notification
+        self._run_retention_period: str | None = run_retention_period
+        self._upload_status: dict[str, str | float] = {}
         self._id_mapping = {
             file_path.name.split(".")[0]: file_path.read_text()
             for file_path in self._cache_directory.glob("server_ids/*.txt")
@@ -96,7 +114,14 @@ class Sender:
             name=f"sender_upload_{_time_stamp}",
             folder="/sender",
             notification=self._run_notification,
+            description="Simvue sender upload session.",
+            retention_period=self._run_retention_period,
             timeout=None,
+            metadata={
+                f"sender.item_count.{upload_object.object_type}": _obj_count
+                for upload_object in UPLOAD_ACTION_ORDER
+                if (_obj_count := upload_object.count(self._cache_directory)) > 0
+            },
             no_color=True,
         )
         _ = _run.config(suppress_errors=True, enable_emission_metrics=False)
@@ -107,14 +132,23 @@ class Sender:
             trigger_abort=False,
         )
 
+        _run.upload_count = 0
+
         return _run
 
     @pydantic.validate_call
     def upload(self, objects_to_upload: list[UploadItem] | None = None) -> None:
-        """Upload objects to server."""
+        """Upload objects to server.
+
+        Parameters
+        ----------
+        objects_to_upload : list[str]
+            Types of objects to upload, by default uploads all types of objects present in cache
+        """
         self._lock()
 
         _monitor_run = self._initialise_monitor_run()
+        self._upload_status = {}
 
         for action in UPLOAD_ACTION_ORDER:
             if objects_to_upload and action.object_type not in objects_to_upload:
@@ -130,8 +164,10 @@ class Sender:
                 thread_lock=self._thread_lock,
                 throw_exceptions=self._throw_exceptions,
                 retry_failed=self._retry_failed_uploads,
-                single_thread_limit=self._threading_threshold,
+                threading_threshold=self._threading_threshold,
                 max_thread_workers=self._max_workers,
                 simvue_monitor_run=_monitor_run,
+                upload_status=self._upload_status,
             )
+        _monitor_run.close()
         self._release()